feat(packages/sui-pde): add decide callback
