+++ packages/sui-pde/src/hooks/common/platformStrategies.js
@@ -6,8 +6,8 @@ const getServerStrategy = () => ({
   getVariation: ({pde, experimentName, attributes, adapterId}) => {
     return pde.getVariation({pde, name: experimentName, attributes, adapterId})
   },
-  decide: ({pde, name, attributes, adapterId}) => {
-    return pde.decide({pde, name, attributes, adapterId})
+  decide: ({pde, name, attributes, adapterId, isEventDisabled}) => {
+    return pde.decide({pde, name, attributes, adapterId, isEventDisabled})
   },
   trackExperiment: () => {},
   getForcedValue: ({key, queryString}) => {
@@ -30,8 +30,8 @@ const getBrowserStrategy = ({customTrackExperimentViewed, cache}) => ({
 
     return variationName
   },
-  decide: ({pde, name, attributes, adapterId}) => {
-    return pde.decide({pde, name, attributes, adapterId})
+  decide: ({pde, name, attributes, adapterId, isEventDisabled}) => {
+    return pde.decide({pde, name, attributes, adapterId, isEventDisabled})
   },
   trackExperiment: ({variationName, experimentName}) => {
     if (customTrackExperimentViewed) {
